requirements.txt
dash
dash-bootstrap-components
pandas
sqlalchemy
psycopg2-binary
plotly
python-dotenv